pyproject.toml
[tool.poetry]
name = "huobi-spot-order-notification"
version = "0.1.0"
description = ""
authors = ["kuznetsov-m"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.13"
huobi-client = {url = "https://github.com/kuznetsov-m-forks/huobi_Python/archive/refs/tags/v2.0.1.tar.gz"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
